requirements.txt
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
flake8==3.9.2
flake8-broken-line==0.3.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
idna==3.3
iniconfig==1.1.1
isort==5.9.3
mccabe==0.6.1
packaging==21.0
pep8-naming==0.12.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.2.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
testfixtures==6.18.3
toml==0.10.2
urllib3==1.26.7
